Create Elon package
