+++ src/main/java/Elon/ElonException.java
@@ -1,3 +1,5 @@
+package Elon;
+
 public class ElonException extends Exception{
     ElonException(String message) {
         super(message);
