+++ src/main/java/Elon/Event.java
@@ -1,3 +1,5 @@
+package Elon;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 public class Event extends Task {
