+++ src/main/java/Elon/Parser.java
@@ -1,3 +1,5 @@
+package Elon;
+
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Scanner;
